// SISPRO/yitpro-react/src/model/Actividad.model.ts
import type { CatalogoGeneralModel } from "./CatalogoGeneral.model";

export interface ActividadesModel {
    idActividad: number;
    seleccionado: boolean | null;
    idActividadStr: string;
    idUsuarioAsignado: number | null;
    bRSE: string;
    bR: string;
    backlog: boolean | null;
    descripcion: string;
    criterioAceptacion: string;
    documentoRef: string;
    planificada: number | null;
    idIteracion: number | null;
    horasFacturables: number | null;
    estatus: string;
    prioridad: number;
    comentariosFinales: string;
    horasAsignadas: number | null;
    tiempoEjecucion: number | null;
    horasFinales: number;
    idProyecto: number;
    tipoActividadId: number;
    idUsuarioResponsable: number | null;
    clasificacionId: number | null;
    idActividadRef: number;
    idActividadRefStr: string;
    fechaSolicitado: string | null;
    fechaInicio: string | null;
    fechaTermino: string | null;
    fechaRevision: string | null;
    fechaCierre: string | null;
    fechaFin: string | null;
    motivoRechazoId: number | null;
    descripcionRechazo: string;
    evidenciaRechazo: string;
    idUCreo: number;
    fechaCreo: string;
    idUMod: number | null;
    fechaMod: string | null;
    retrabajo: boolean;
    estatusStr: string;
    tipoActividadStr: string;
    clasificacionStr: string;
    prioridadStr: string;
    asignadoStr: string;
    asignadoPath: string;
    responsableStr: string;
    responsablePath: string;
    proyectoStr: string;
    correoResponsable: string;
    motivoRechazoStr: string;
    usuarioRechazoStr: string;
    fechaRechazo: string | null;
    correoLider: string;
    claveTipoActividad: string;
    claveClasificacionActividad: string;
    claveUsuario: string;
    claveProyecto: string;
    sprint: string;
    estatusCte: string;
    estatusCteStr: string;
    tipoId: number | null;
    tipoNombre: string;
    tipoUrl: string;
    comentario: string;
    idUsuario: number;
    tracking: never[];//ActividadTrackingModel[];
    comentarios: never[];//ActividadComentarioModel[];
    archivos: never[];//ActividadArchivoModel[];
    trabajos: never[];//ActividadTrabajoModel[];
    dependencias: never[];//ActividadesModel[];
    actividadLog: never[];//ActividadComentarioModel[];
    validaciones: never[];//ActividadesValidacionModel[];
    issues: never[];//ProyectoIssueModel[];
    progreso: number | null;
    progresoStr: string;
    comentariosTotal: number;
    archivosTotal: number;
    tipo: number;
    pSP: number;
    idActividadDependencia: number;
    idActividadD: number;
    dependenciasA: number;
    dependenciasT: number;
    avanceDependencia: number;
    fechaActual: string;
    totalComentarios: number;
    totalArchivos: number;
    totalDependencias: number;
    totalLog: number;
    totalTiempos: number;
    totalValidaciones: number;
    critico: boolean;
    prioridadId: number | null;
    idListaRevision: number | null;
    esPeerReview: boolean;
    actividadRepositorioLinks: never[];//ActividadRepositorioModel[];
    proyectoRepositorio: CatalogoGeneralModel[];
    icono: string;
    termino: string;
    totalIssues: number;
    fInicio: string;
    fFin: string;
    jerarquia: number;
    idActividadR1: number;
    idActividadR2: number;
    idUModStr: string;
    fechaLiberacion: string | null;
    correoAsignado: string;
    avanceUX: string;
    avanceBugs: string;
    avanceDev: string;
    avanceCalidad: string;
    avanceImp: string;
    idCicloCaso: number;
    objetivo: number;
    performance: number;
    puntos: number | null;
    idWorkFlow: number;
    workFlow: string;
    colorW: string;
    hU: string;
    colorTexto: string;
    claveResponsable: string;
    idActividadRelacionada: number | null;
    orden: number;
}